updated 4/24/204 at 2115 hours - code cleanup and fixed clear data alert
Cleaned out all the commented code and commented out dead pages with no apparent use for testing of deletion.

A new bug arose in the lack of alert for hitting clear data button for medicalschooldetails as well as Internship, residency and fellowship.  Employment details did not have either save data alert or clear data alert.  This was due to inadvertant deletion of code for the alerts on detail view pages.  The board certification, state medical license, hospital affiliations and private tab had no problems with the save and clear data alerts